Move helpers.go to cli/commands
