+++ cli/commands/helpers.go
@@ -1,4 +1,4 @@
-package termio
+package commands
 
 import (
 	"errors"
@@ -12,9 +12,9 @@ import (
 	"github.com/raedahgroup/godcr/walletrpcclient"
 )
 
-// GetSendSourceAccount fetches the account number for an account selected by the user.
+// getSendSourceAccount fetches the account number for an account selected by the user.
 // If there is only one account available, it returns the account number for that account.
-func GetSendSourceAccount(c *walletrpcclient.Client) (uint32, error) {
+func getSendSourceAccount(c *walletrpcclient.Client) (uint32, error) {
 	var selection int
 	var err error
 	// get send  accounts
@@ -54,8 +54,8 @@ func GetSendSourceAccount(c *walletrpcclient.Client) (uint32, error) {
 	return accounts[selection].AccountNumber, nil
 }
 
-// GetSendDestinationAddress fetches the destination address to send DCRs to from the user.
-func GetSendDestinationAddress(c *walletrpcclient.Client) (string, error) {
+// getSendDestinationAddress fetches the destination address to send DCRs to from the user.
+func getSendDestinationAddress(c *walletrpcclient.Client) (string, error) {
 	validateAddressInput := func(address string) error {
 		isValid, err := c.IsAddressValid(address)
 		if err != nil {
@@ -77,8 +77,8 @@ func GetSendDestinationAddress(c *walletrpcclient.Client) (string, error) {
 	return address, nil
 }
 
-// GetSendAmount fetches the amout of DCRs to send from the user.
-func GetSendAmount() (float64, error) {
+// getSendAmount fetches the amout of DCRs to send from the user.
+func getSendAmount() (float64, error) {
 	var amount float64
 	var err error
 
@@ -99,8 +99,8 @@ func GetSendAmount() (float64, error) {
 	return amount, nil
 }
 
-// GetWalletPassphrase fetches the user's wallet passphrase from the user.
-func GetWalletPassphrase() (string, error) {
+// getWalletPassphrase fetches the user's wallet passphrase from the user.
+func getWalletPassphrase() (string, error) {
 	result, err := terminalprompt.RequestInputSecure("Wallet Passphrase", terminalprompt.EmptyValidator)
 	if err != nil {
 		return "", fmt.Errorf("error receiving input: %s", err.Error())
@@ -108,8 +108,8 @@ func GetWalletPassphrase() (string, error) {
 	return result, nil
 }
 
-// GetUtxosForNewTransaction fetches unspent transaction outputs to be used in a transaction.
-func GetUtxosForNewTransaction(utxos []*walletrpcclient.UnspentOutputsResult, sendAmount float64) ([]string, error) {
+// getUtxosForNewTransaction fetches unspent transaction outputs to be used in a transaction.
+func getUtxosForNewTransaction(utxos []*walletrpcclient.UnspentOutputsResult, sendAmount float64) ([]string, error) {
 	var selectedUtxos []string
 	var err error
